Move joinside struct to common
