+++ datafusion/physical-plan/src/joins/symmetric_hash_join.rs
@@ -43,7 +43,7 @@ use crate::joins::hash_join_utils::{
 use crate::joins::utils::{
     build_batch_from_indices, build_join_schema, check_join_is_valid,
     combine_join_equivalence_properties, partitioned_join_output_partitioning,
-    prepare_sorted_exprs, ColumnIndex, JoinFilter, JoinOn, JoinSide,
+    prepare_sorted_exprs, ColumnIndex, JoinFilter, JoinOn,
 };
 use crate::{
     expressions::{Column, PhysicalSortExpr},
@@ -58,7 +58,9 @@ use arrow::compute::concat_batches;
 use arrow::datatypes::{Schema, SchemaRef};
 use arrow::record_batch::RecordBatch;
 use datafusion_common::utils::bisect;
-use datafusion_common::{internal_err, plan_err, DataFusionError, JoinType, Result};
+use datafusion_common::{
+    internal_err, plan_err, DataFusionError, JoinSide, JoinType, Result,
+};
 use datafusion_execution::memory_pool::MemoryConsumer;
 use datafusion_execution::TaskContext;
 use datafusion_physical_expr::intervals::ExprIntervalGraph;
